Refactor votemap cache migrations and enhance Redis database handling
- Updated entrypoint to use the new cache_migrations module.
- Introduced main.py for centralized migration logic.
- Added redis_databases.py for Redis client management across databases.
- Implemented votemap.py for handling votemap state migrations.
- Cleaned up maps_history.py by removing redundant functions.
- Added tests for votemap cache migration to ensure data integrity during upgrades.

This refactor improves the structure and maintainability of the votemap migration process.
