+++ rcon/cache_migrations/maps_history.py
@@ -26,6 +26,10 @@ from typing import Any, cast
 import orjson
 import redis
 
+from rcon.cache_migrations.redis_databases import (
+    populated_database_numbers as _populated_database_numbers,
+    redis_client_for_database as _redis_client_for_database,
+)
 from rcon.types import MapInfo
 
 logger = logging.getLogger(__name__)
@@ -232,22 +236,6 @@ def migrate_maps_history(history: Any, max_retries: int = 3) -> int:
                 raise
 
     return 0
-
-
-def _redis_client_for_database(client: redis.Redis, database: int) -> redis.Redis:
-    connection_kwargs = dict(client.connection_pool.connection_kwargs)
-    connection_kwargs["db"] = database
-    return redis.Redis(connection_pool=redis.ConnectionPool(**connection_kwargs))
-
-
-def _populated_database_numbers(client: redis.Redis) -> list[int]:
-    databases = {int(client.connection_pool.connection_kwargs.get("db", 0))}
-    for name in client.info("keyspace"):
-        if isinstance(name, bytes):
-            name = name.decode()
-        if name.startswith("db") and name[2:].isdigit():
-            databases.add(int(name[2:]))
-    return sorted(databases)
 
 
 def migrate_all_maps_histories(redis_url: str) -> tuple[int, int]:
