+++ rcon/vote_map/service.py
@@ -2,13 +2,11 @@ from dataclasses import dataclass
 from enum import Enum, IntFlag
 import functools
 import logging
-import pickle
 import random
-from rcon.maps import Layer, get_all_layers_by_map
 import re
 from datetime import UTC, datetime, timedelta
 from functools import partial
-from typing import Any, Dict, List, Optional, Set, Tuple, cast
+from typing import Dict, List, Optional, Tuple
 from collections.abc import Iterable, Callable
 from collections import Counter
 
@@ -16,7 +14,7 @@ from sqlalchemy import and_
 from sqlalchemy.exc import MultipleResultsFound, NoResultFound
 
 from rcon import maps
-from rcon.cache_utils import get_redis_client, ttl_cache
+from rcon.cache_utils import ttl_cache
 from rcon.connection import HLLCommandError
 from rcon.maps import (
     Environment,
@@ -32,8 +30,6 @@ from rcon.types import (
     MapInfo,
     PlayerProfileType,
     StructuredLogLineWithMetaData,
-    VoteMapHistory,
-    VoteMapHistoryResult,
     VoteMapStatus,
     VoteMapVote,
     VoteMapMapResult,
@@ -44,7 +40,21 @@ from rcon.user_config.vote_map import (
     DefaultMethods,
     VoteMapUserConfig,
 )
-from rcon.utils import FixedLenList, MapsHistory
+from rcon.utils import MapsHistory
+
+from .exceptions import (
+    InvalidMapParam,
+    InvalidVoteError,
+    PlayerChoiceNotAllowed,
+    PlayerNotFound,
+    PlayerVoteNotAllowed,
+    RestrictiveFilterError,
+    SelectionLimitExceeded,
+    VoteMapException,
+    VoteMapNoInitialised,
+)
+from .selection import MapSelectionBuilder, MapSelectionCriteria
+from .state import VotemapState
 
 logger = logging.getLogger(__name__)
 
@@ -334,7 +344,9 @@ class VoteMap:
                 f"Cannot add map {map.pretty_name} to votemap selection. Max selection limit {self.SELECTION_LIMIT} would be exceeded."
             )
         if map in selection:
-            raise VoteMapException(f"Map {map.pretty_name} is already in the selection.")
+            raise VoteMapException(
+                f"Map {map.pretty_name} is already in the selection."
+            )
 
         if index is not None:
             selection.insert(index, map)
@@ -657,7 +669,6 @@ class VoteMap:
                 player_choice=self.get_player_choice(),
             )
             reminder_message = instructions.format(map_selection=map_selection)
-            
 
         for name, player_id in players_to_remind:
             try:
@@ -1433,463 +1444,3 @@ class VoteMapCommandHandler:
             return params.pop(0)
         except IndexError:
             return default
-
-
-class VotemapState:
-    VERSION = "votemap:version"
-    LATEST_REMINDER = "votemap:reminder"
-    MAP_WHITELIST = "votemap:whitelist"
-    MAP_SELECTION = "votemap:selection"
-    VOTES = "votemap:votes"
-    ADMIN_NEXT_MAP = "votemap:admin-next-map"
-    PLAYER_CHOICE = "votemap:player-choice"
-    NEXT_MAP = "votemap:next-map"
-    RESULT_HISTORY = "votemap:result-history"
-
-    def __init__(self) -> None:
-        self.version = 0.2
-        self.client = get_redis_client()
-        self.results = FixedLenList[list[VoteMapHistory]](self.RESULT_HISTORY)
-        if self.get_version() != self.version:
-            self.delete_last_reminder_time()
-            self.delete_next_map()
-            self.delete_admin_next_map()
-            self.delete_player_choice()
-            self.delete_selection()
-            self.delete_votes()
-            self.set_version(self.version)
-
-    ###
-    # VERSION
-    ###
-    def get_version(self):
-        version = cast(bytes | None, self.client.get(self.VERSION))
-        if version is None:
-            return None
-        return float(version.decode())
-
-    def set_version(self, version: float):
-        self.client.set(self.VERSION, version)
-
-    ###
-    # PLAYER CHOICE
-    ###
-    def get_player_choice(self) -> Dict[str, str] | None:
-        raw = cast(Dict[bytes, bytes], self.client.hgetall(self.PLAYER_CHOICE))
-        if not raw:
-            return None
-        return {k.decode(): v.decode() for k, v in raw.items()}
-
-    def set_player_choice(self, player_id: str, player_name: str):
-        self.client.hset(
-            self.PLAYER_CHOICE,
-            mapping={"player_name": player_name, "player_id": player_id},
-        )
-
-    def delete_player_choice(self):
-        self.client.delete(self.PLAYER_CHOICE)
-
-    ###
-    # ADMIN NEXT MAP
-    # When the admin wants to guarantee the next map
-    # That practically pauses the map voting
-    ###
-    def get_admin_next_map(self) -> Layer | None:
-        raw = cast(bytes | None, self.client.hget(self.ADMIN_NEXT_MAP, "map_name"))
-        if not raw:
-            return None
-        return maps.parse_layer(raw.decode())
-
-    def set_admin_next_map(self, map: Layer):
-        self.client.hset(
-            self.ADMIN_NEXT_MAP,
-            key="map_name",
-            value=map.id,
-        )
-
-    def delete_admin_next_map(self):
-        self.client.delete(self.ADMIN_NEXT_MAP)
-
-    ###
-    # LAST REMINDER TIME
-    ###
-    def get_last_reminder_time(self) -> datetime | None:
-        as_date: datetime | None = None
-        res = self.client.get(self.LATEST_REMINDER)
-        if res is not None:
-            as_date = pickle.loads(res)  # type: ignore
-        return as_date
-
-    def set_last_reminder_time(self, the_time: datetime) -> None:
-        self.client.set(self.LATEST_REMINDER, pickle.dumps(the_time))
-
-    def delete_last_reminder_time(self) -> None:
-        self.client.delete(self.LATEST_REMINDER)
-
-    ###
-    # VOTES
-    ###
-    def get_votes(self) -> list[VoteMapVote]:
-        raw_votes = cast(dict[bytes, bytes], self.client.hgetall(self.VOTES))
-        return [pickle.loads(vote) for _, vote in raw_votes.items()]
-
-    def delete_votes(self):
-        self.client.delete(self.VOTES)
-
-    def get_vote(self, player_id: str) -> VoteMapVote | None:
-        vote = cast(bytes | None, self.client.hget(self.VOTES, player_id))
-        if not vote:
-            return None
-        return pickle.loads(vote)
-
-    def add_vote(
-        self, player_id: str, player_name: str, map: Layer, vote_count: int = 1
-    ):
-        vote = pickle.dumps(
-            {
-                "player_id": player_id,
-                "player_name": player_name,
-                "map_id": map.id,
-                "vote_count": vote_count,
-            }
-        )
-        self.client.hset(self.VOTES, player_id, cast(Any, vote))
-
-    def delete_vote(self, player_id):
-        self.client.hdel(self.VOTES, player_id)
-
-    ###
-    # MAP WHITELIST
-    ###
-    def get_whitelist(self) -> list[Layer]:
-        raw = cast(Set[bytes], self.client.smembers(self.MAP_WHITELIST))
-        uniques = {item.decode() for item in raw}
-        return [maps.parse_layer(map_id) for map_id in uniques]
-
-    def add_map_to_whitelist(self, map: Layer):
-        self.client.sadd(self.MAP_WHITELIST, map.id)
-
-    def remove_map_from_whitelist(self, map: Layer):
-        self.client.srem(self.MAP_WHITELIST, map.id)
-
-    def set_whitelist(self, maps: Iterable[Layer]):
-        self.client.delete(self.MAP_WHITELIST)
-        self.client.sadd(self.MAP_WHITELIST, *[map.id for map in maps])
-
-    ###
-    # MAP SELECTION
-    ###
-    def get_selection(self) -> list[Layer]:
-        # Get all map ids in order
-        raw = cast(Set[bytes], self.client.zrange(self.MAP_SELECTION, 0, -1))
-        return [maps.parse_layer(item.decode()) for item in raw]
-
-    def set_selection(self, maps: Iterable[Layer]):
-        self.delete_selection()
-        # Add each map_id with its index as score
-        for idx, map in enumerate(maps):
-            map_id = map.id
-            self.client.zadd(self.MAP_SELECTION, {map_id: idx})
-
-    def remove_map_from_selection(self, map: Layer):
-        self.client.zrem(self.MAP_SELECTION, map.id)
-
-    def delete_selection(self):
-        self.client.delete(self.MAP_SELECTION)
-
-    ###
-    # NEXT MAP
-    ###
-    def get_next_map(self) -> Layer | None:
-        next_map = cast(bytes | None, self.client.get(self.NEXT_MAP))
-        if next_map is None:
-            return None
-        return maps.parse_layer(next_map.decode())
-
-    def set_next_map(self, map: Layer):
-        self.client.set(self.NEXT_MAP, map.id)
-
-    def delete_next_map(self):
-        self.client.delete(self.NEXT_MAP)
-
-    ###
-    # RESULT HISTORY
-    ###
-    def get_results(self):
-        history = FixedLenList[list[VoteMapHistory]](self.RESULT_HISTORY)
-        return [
-            {
-                "map": maps.parse_layer(x["map_id"]),
-                "ts": x["ts"],
-                "results": [
-                    {
-                        "map": maps.parse_layer(r["map_id"]),
-                        "votes_count": r["votes_count"],
-                    }
-                    for r in x["results"]
-                ],
-            }
-            for x in history
-        ]
-
-    def add_result(self, status: VoteMapStatus, map_id: str, ts: int):
-        results = [
-            VoteMapHistoryResult(map_id=str(res["map"]), votes_count=res["votes_count"])
-            for res in status["results"]
-        ]
-        self.results.add(VoteMapHistory(map_id=map_id, ts=ts, results=results))
-
-
-@dataclass(frozen=True)
-class MapSelectionCriteria:
-    """User-configured rules and context used to build a votemap selection."""
-
-    maps_history: list[Layer]
-    current_map: Layer
-    allowed_maps: set[Layer]
-    number_last_played_to_exclude: int
-    num_warfare_options: int
-    num_offensive_options: int
-    num_skirmish_control_options: int
-    consider_offensive_same_map: bool
-    consider_skirmishes_as_same_map: bool
-    consider_environment_as_same_map: bool
-    allow_multiple_maps_with_same_environment: bool
-    allow_consecutive_offensives: bool
-    allow_consecutive_offensives_opposite_sides: bool
-    allow_consecutive_skirmishes: bool
-
-
-class MapSelectionBuilder:
-    """Builds a randomized votemap selection by filtering and sampling map layers."""
-
-    def build(self, criteria: MapSelectionCriteria) -> list[Layer]:
-        """Return a filtered, randomized map list for players to vote on."""
-        candidates = self._build_candidates(criteria)
-        selection = self._sample_selection(candidates, criteria)
-
-        if not selection:
-            logger.error("No maps can be suggested with the given parameters.")
-            raise RestrictiveFilterError("Unable to suggest map")
-
-        logger.info("Suggestion %s", [m.pretty_name for m in selection])
-        return selection
-
-    def _log_candidates(self, step: str, candidates: set[Layer]) -> None:
-        """Log the candidate pool remaining after a filter step."""
-        logger.info(
-            "%s - remaining maps to suggest from: %s",
-            step,
-            [m.pretty_name for m in candidates],
-        )
-
-    def _history_exclusions(self, criteria: MapSelectionCriteria) -> set[Layer]:
-        """Return layers excluded because they were played too recently."""
-        if criteria.number_last_played_to_exclude <= 0:
-            return set()
-
-        last_n_maps = criteria.maps_history[: criteria.number_last_played_to_exclude]
-        if criteria.consider_environment_as_same_map:
-            excluded = {
-                variant
-                for layer in last_n_maps
-                for variant in get_all_layers_by_map(
-                    layer.map, layer.game_mode, layer.attackers
-                )
-            }
-        else:
-            excluded = set(last_n_maps)
-
-        logger.info(
-            "Excluding last %s played maps: %s",
-            criteria.number_last_played_to_exclude,
-            [m.pretty_name for m in excluded],
-        )
-        return excluded
-
-    def _exclude_same_map_variants(
-        self,
-        candidates: set[Layer],
-        history_excluded: set[Layer],
-        criteria: MapSelectionCriteria,
-    ) -> set[Layer]:
-        """Remove all modes on a map when offensive/skirmish should count as the same map."""
-        if not (
-            criteria.consider_offensive_same_map
-            or criteria.consider_skirmishes_as_same_map
-        ):
-            return candidates
-
-        excluded_map_ids = {layer.map for layer in history_excluded}
-        logger.info(
-            "Considering offensive/skirmish mode as same map, excluding %s",
-            excluded_map_ids,
-        )
-        return {layer for layer in candidates if layer.map not in excluded_map_ids}
-
-    def _exclude_consecutive_mode_maps(
-        self, candidates: set[Layer], criteria: MapSelectionCriteria
-    ) -> set[Layer]:
-        """Remove offensive or skirmish layers when the current map disallows back-to-back modes."""
-        current_map = criteria.current_map
-
-        if (
-            not criteria.allow_consecutive_offensives
-            and current_map.game_mode == GameMode.OFFENSIVE
-        ):
-            logger.info(
-                "Current map %s is offensive. Excluding all offensives from suggestions",
-                current_map,
-            )
-            candidates = {
-                layer for layer in candidates if layer.game_mode != GameMode.OFFENSIVE
-            }
-
-        if not criteria.allow_consecutive_skirmishes and current_map.game_mode in (
-            GameMode.SKIRMISH,
-            GameMode.PHASED,
-            GameMode.MAJORITY,
-        ):
-            logger.info(
-                "Current map %s is skirmish. Excluding all skirmishes from suggestions",
-                current_map,
-            )
-            candidates = {
-                layer for layer in candidates if layer.game_mode != GameMode.SKIRMISH
-            }
-
-        return candidates
-
-    def _exclude_opposite_side_offensives(
-        self, candidates: set[Layer], criteria: MapSelectionCriteria
-    ) -> set[Layer]:
-        """Remove offensives that would make the same team defend twice in a row."""
-        current_map = criteria.current_map
-        if (
-            criteria.allow_consecutive_offensives_opposite_sides
-            or current_map.game_mode != GameMode.OFFENSIVE
-        ):
-            return candidates
-
-        if current_map.attackers:
-            logger.info(
-                "Not allowing consecutive offensive with opposite side: %s",
-                maps.get_opposite_side(current_map.attackers),
-            )
-        return {
-            layer
-            for layer in candidates
-            if layer.opposite_side != current_map.attackers
-        }
-
-    def _build_candidates(self, criteria: MapSelectionCriteria) -> set[Layer]:
-        """Apply all exclusion rules to the whitelist and return the remaining candidates."""
-        history_excluded = self._history_exclusions(criteria)
-        candidates = criteria.allowed_maps - history_excluded
-        self._log_candidates("After excluding recently played maps", candidates)
-
-        candidates = self._exclude_same_map_variants(
-            candidates, history_excluded, criteria
-        )
-        self._log_candidates("After excluding same-map variants", candidates)
-
-        candidates = self._exclude_consecutive_mode_maps(candidates, criteria)
-        self._log_candidates("After excluding consecutive mode maps", candidates)
-
-        candidates = self._exclude_opposite_side_offensives(candidates, criteria)
-        self._log_candidates("After excluding opposite-side offensives", candidates)
-
-        return candidates
-
-    def _get_random_map_selection(
-        self,
-        maps_pool: list[Layer],
-        nb_to_return: int,
-        allow_multiple_maps_with_same_environment: bool,
-    ) -> list[Layer]:
-        """Pick up to `nb_to_return` random layers, optionally limiting to one environment per map/mode/team."""
-        try:
-            if nb_to_return > 0 and len(maps_pool) < nb_to_return:
-                nb_to_return = len(maps_pool)
-
-            if not allow_multiple_maps_with_same_environment:
-                selected_maps: list[Layer] = []
-                pool = list(maps_pool)
-                while len(pool) > 0 and len(selected_maps) < nb_to_return:
-                    pick = random.choice(pool)
-                    selected_maps.append(pick)
-                    pool = [
-                        layer
-                        for layer in pool
-                        if not (
-                            layer.map == pick.map
-                            and layer.game_mode == pick.game_mode
-                            and layer.attackers == pick.attackers
-                        )
-                    ]
-                return selected_maps
-
-            return random.sample(maps_pool, k=nb_to_return)
-        except (IndexError, ValueError):
-            return []
-
-    def _sample_selection(
-        self, candidates: set[Layer], criteria: MapSelectionCriteria
-    ) -> list[Layer]:
-        """Randomly sample warfare, offensive, and skirmish options from the filtered candidates."""
-        categorized_maps = categorize_maps(candidates)
-        return (
-            self._get_random_map_selection(
-                categorized_maps[GameMode.OFFENSIVE],
-                criteria.num_offensive_options,
-                criteria.allow_multiple_maps_with_same_environment,
-            )
-            + self._get_random_map_selection(
-                categorized_maps[GameMode.WARFARE],
-                criteria.num_warfare_options,
-                criteria.allow_multiple_maps_with_same_environment,
-            )
-            + self._get_random_map_selection(
-                categorized_maps[GameMode.SKIRMISH],
-                criteria.num_skirmish_control_options,
-                criteria.allow_multiple_maps_with_same_environment,
-            )
-        )
-
-
-class RestrictiveFilterError(Exception):
-    pass
-
-
-class VoteMapException(Exception):
-    pass
-
-
-class SelectionLimitExceeded(VoteMapException):
-    pass
-
-
-class InvalidVoteError(VoteMapException):
-    pass
-
-
-class VoteMapNoInitialised(VoteMapException):
-    pass
-
-
-class PlayerNotFound(Exception):
-    pass
-
-
-class PlayerVoteNotAllowed(VoteMapException):
-    def __init__(self, message="Player is not allowed to vote map"):
-        super().__init__(message)
-
-
-class PlayerChoiceNotAllowed(VoteMapException):
-    def __init__(self, message="Player does not have any of required flags."):
-        super().__init__(message)
-
-
-class InvalidMapParam(VoteMapException):
-    pass
